Fix rendering palettized gdal sources.
Also, for gdal sources with a red, green, and blue channel, render them
using the specified maximum and minimum values rather than the default
which uses 0-255.

It would be a nice extension to allow a style of band: `{r:#, g:#, b:#
[a:#]}` rather than just an integer for the band.

We should probably default the min and max of the band in any case
unless the style specifies otherwise.