Use the Girder FUSE plugin.
For files on non-filesystem assetstores, use the Girder FUSE plugin, if
it is enabled and configured appropriately.  This does no harm if the
plugin isn't present, turned off, or unconfigured.

If a tile source sets the `mayHaveAdjacentFiles` flag, then the fuse
path will be used in preference to the direct path.  This allows tile
sources to work where there are multiple files associated by a common
base name (such as shapefiles or vrt files with relative paths).

python-mapnik hangs with FUSE.  See
https://github.com/mapnik/python-mapnik/issues/182.